||dW||_F in fp64: the fp32 path was ~0.45% low, ratios unaffected
Chasing a 0.4% gap between the low-rank trace identity and the direct norm
showed the identity was right and the DIRECT path wrong: in fp64 the two
agree to 7e-14, while fp32-direct sits 4.5e-3 below the fp64 reference. It
sums ~58M squared terms in fp32, which is where the error accumulates.

What this does and does not change. A2 (repro/released norm ratio), A3
(Spearman over per-module norms) and the 6c cross-term share are all ratios
or ranks of two norms computed the same way, so the bias cancels and every
reported value stands. Only absolute norms move: 9.09 -> ~9.13.

+++ scripts/gate_weight_checks.py
@@ -46,13 +46,18 @@ def module_dW_norms(adapter_dir: str | Path) -> tuple[dict[str, float], dict]:
         m = KEY.match(k)
         if not m:
             continue
-        pairs.setdefault(m.group(1), {})[m.group(2)] = v.to(torch.float32)
+        pairs.setdefault(m.group(1), {})[m.group(2)] = v.to(torch.float64)
     out = {}
     for mod, ab in pairs.items():
         if "A" not in ab or "B" not in ab:
             continue
         # ||dW||_F with dW = scaling * B @ A, formed one module at a time
-        out[mod] = float(torch.linalg.matrix_norm(scaling * (ab["B"] @ ab["A"]), ord="fro"))
+        # fp64: ||B@A||_F in fp32 sums ~58M squared terms and comes out ~0.45% low
+        # (checked against an fp64 reference and the exact low-rank trace identity).
+        # Ratios of two such norms are unaffected -- the bias cancels -- so A2/A3 and
+        # the 6c cross-term share as already reported stand; absolute norms shift ~0.45%.
+        out[mod] = float(torch.linalg.matrix_norm(
+            scaling * (ab["B"].double() @ ab["A"].double()), ord="fro"))
     return out, cfg
